Extract SnideSelectCard to its own file, byte-for-byte (#2322)
The pure move, so the repaint that follows has nothing structural in it.
`SnideSelectCard` leaves the 595-line `FactionSelectCard.tsx` for
`components/selectCard/SnideSelectCard.tsx`, matching how
`components/taskCard/` is laid out; the dispatcher keeps only the shared
types, `LEGACY_SLUG` and `FactionSelectCard` itself. The manifest row is
unchanged in shape — `lazyArchetype(() => import(...))` now points at the
new module and drops the `.then((m) => ({ default: m.SnideSelectCard }))`
unwrap, because the new file default-exports it the way every task card
does.

PROOF THE DIRECTORY IS UNCHANGED. `renderToStaticMarkup` of
`<FactionSelectCard faction="snide" />` across all six state x members
combinations is byte-identical before and after this commit (24,502 bytes,
`diff` clean). The probe was a throwaway; the standing guards are
`factionSelectCardDispatch.test.tsx` and `snideSigilIsTheAnarchyA.test.tsx`,
both green.

THE DISPATCHER'S "the tiles are dark" IS DELETED, not carried into the new
file. It was already false — Coven's tile is pink and UA's is cream — and it
is the premise that let a dark-only register look correct on a surface that
flips (#2321).

`npm run lint` is RED at this commit alone, by design: the tile's
`1px solid #000` hairline is allowlisted on
`.eslint-legacy-raw-colours.txt` by PATH, and the path just changed. The
next commit retires the hairline and takes the line off the list, rather
than adding a new file to a list that may only shrink.
