+++ frontend/src/factions/snide.ts
@@ -29,7 +29,7 @@ const SnidePraxisCard = lazyArchetype(() => import('../components/praxisCard/des
 const SnideScoreStamp = lazyArchetype(() => import('../components/praxisCard/scoreStamp/SnideScoreStamp'))
 const SnideSeal = lazyArchetype(() => import('../components/metataskSeal/skins/SnideSeal'))
 const SnideSigil = lazyArchetype(() => import('../components/sigil/SnideSigil').then((m) => ({ default: m.SnideSigil })))
-const SnideSelectCard = lazyArchetype(() => import('../components/selectCard/FactionSelectCard').then((m) => ({ default: m.SnideSelectCard })))
+const SnideSelectCard = lazyArchetype(() => import('../components/selectCard/SnideSelectCard'))
 
 export const SNIDE_MANIFEST: FactionManifest = {
   slug: 'snide',
